Fix tests for LOAD/HEADER, URL for console error problem
Due to editing files on a borrowed machine that didn't have git, the
manual transcription of a problem case URL was mistyped on the
machine that did have git.  This fixes that URL.

Also updates two tests of LOAD that didn't get properly adjusted
when the multiple return value change was made.