Attempt to restore Ctrl-C cancellation behavior
Calls into the libRebol API had been able to be halted via Ctrl-C.
This would cause the same behavior as raising an error would...using
longjmp (but could be changed to use C++ exceptions or JavaScript
exceptions).  It meant that there would have to be some mechanism set
up for catching it, or the program would crash.

A key difference between Ctrl-C and ordinary errors is that they can
occur at any moment.  This meant that as the console itself became
driven by more and more usermode code (e.g. processing individual
keystrokes) the very logic of handling Ctrl-C itself could be
interrupted.

This led to the need to differentiate interruptible API calls from
non-interruptible ones.  As a first step in this idea, this takes
the main routine that was used in the console to run code typed
in by the user (rebQuote()) and made an interruptible version
(rebQuoteInterruptible()).  All other APIs were transitioned to a
default of being not interruptible.

It was enough to deal with the broken nature of Ctrl-C for the moment--
though more review is needed.