Restore sigmask on errors (further improves Ctrl-C)
In the fledgling idea of making API calls interruptible or not, there
was a hole in the handling of longjmps.  Even if code would temporarily
disable the halt signal...with intent to put it back later...a longjmp
could go past the API call that was planning to do so.  (Similar issues
exist in POSIX signals when interacting with signal masks and longjmp
for the signals inside the OS itself).

This addresses that.  It also allows you to have interruptible signals
nested within non-interruptible ones (though any interruptible call
must be in a rebRescue to process it at the moment).