Improve redbol emulation of FUNC/FUNCTION args
Historical Rebol and Red did not have their functions use the
refinements as the arguments themselves.  This changed in Ren-C in
order to make frames more uniform--partially enabled because the
availability of NULL meant refinements could have a state that did not
correspond to an "actual value".  This resulted in simplification of
the evaluator, reducing the size of frames, and making it easier to
conceptualize operations like APPLY.

This updates the logic for emulating the old behavior in Redbol.