Don't ensure console position in debug build by default
While trying to diagnose a problem in the key-by-key console code,
an assertion was added that used Windows knowledge of the cursor
position to check to see if the internal index into the "virtual"
terminal buffer matched the idea that Windows had.

This assert is currently easy to break by doing various things, like
resizing the console or using a printf() (that doesn't go through
the smart terminal code, only used in debug bulids).  So it should
not be on unless a bug is being investigated, as it's too annoying
to crash and lose your state due to a visual problem in typing.