Defer items between right soft quote and left quote
There has already been a resolution for left-quoting and right-quoting
operations that directly compete.  The right quoting operation wins
*unless* it is at the tail of a series; an exception which permits
HELP to work on things like DEFAULT.

However, there is another case of contention for things that are
*between* left-quoting and right quoting constructs.  A good example
comes up with lambda, a shorthand for declaring functions and THEN:

    [x] => [print [x]]  ; like func [x] [print [x]]

Conditionals like IF and THEN use quoting on their branches:

    >> if true [print "Hello"]
    Hello

    >> if true '[print "Hello"]
    == [print "Hello"]

But this means that the following construct creates a "quote fight"

    if condition [...] then [x] => [print [x]]

A general rule which seems to apply in the language is that expressions
that quote left do so in order to generate a value, while expressions
that quote right are attempting to consume a value.  In this case,
the generation is of a function--with a different syntax.  So the
THEN defers the calculation to the right, effectively giving:

    if condition [...] then ([x] => [print [x]])