Check environment var for R3_ALWAYS_MALLOC in valgrind
Valgrind can find more problems if you use malloc() than if there
are pooled allocations that it can't tell when conditional jumps
happen on uninitialized values (Address Sanitizer doesn't check this).

This makes the callgrind build check for R3_ALWAYS_MALLOC in the
environment.  A better solution might let you pick it independently
so it didn't interfere with the speed of memory allocation, but it's
extremely negligible.

It also goes ahead and does a "clean" shutdown despite being an
ostensibly "release" build that could leave memory outstanding.  This
interferes with the performance aspect of callgrind if you want
to time the shutdown...but keeps from contaminating the output if
you are doing memcheck in valgrind instead.