Fix Find_Str_In_Str boundaries / skipping
Using Valgrind with R3_ALWAYS_MALLOC exposed that finding one string
inside of another could red bytes outside of the initialized range.
This didn't show up in address sanitizer due to memory pooling--where
the pools looked "used" once they had ever been written to.

While attempting to address that issue, the resulting review of the
code revealed that it was rather flawed.  Though it had gotten
a slight review during the large-scale implementation period of
UTF-8 Everywhere migration, this had not been enough to make the
code solid and robust w.r.t. /SKIP lengths, positive or negative.

This is a rethinking of the mechanics which also updates the
string layer's notion of how SKIP_CHR should work.  The revised
mechanic reads a character at the end of the skip range (forward
or negative) and returns the pointer to that character, and not
past it.  The logic of the Find_Str_In_Str() function shows how
this makes it possible to never be holding a pointer to a byte in an
invalid memory location.