No zero-bytes in ANY-STRING! while allowing NUL char!
One of Ren-C's goals is natural and convenient interoperability with
the C language.  Pursuant to that, an emergent policy decision in Ren-C
has been that zero bytes would not be allowed in the internals of an
ANY-STRING!.  This is to avoid potential pitfalls surrounding asking
a string for its content without also asking it for its length, as
embedded '\0' can wreak havoc in that case.  To help mitigate any
difficulty those working with BINARY! that has text portions inside of
it, several routines make UTF-8 processing via string easier.

This goal has been gradually enforced, but one question was about if
a CHAR! value corresponding 0 should also be illegal.  The general
feeling is that it should be...since you can append CHAR! to a
BINARY! it can be used for things.  Alternate choices like saying
`to char! 0` gives back BLANK! or NULL seem to create confusions and
limitations.

So this goes codifies the <<#"^(00)" is good>> but <<"^(00)" is bad>>
philosophy.  Toward that end, it helps catch more violations by
noticing anywhere that LO_CASE(), UP_CASE(), IS_WHITE(), or
IS_SPACE() were being used on NUL characters--which tightens up
some of the string enumerations where it wasn't clear if you were
looking at a "real" character at that moment or not.

(A similar change happened early on in Ren-C where REB_END cells would
not answer to standard questions like IS_WORD()...a change which
helped clarify code for readability and also caught many bugs.)