Correct overstep in PARSE subroutine
With increased vigilance in the system regarding ANY-STRING! prohbiting
internal '\0' characters, routines that overstep and treat the terminal
'\0' as if it were part of the string raise more asserts.

This corrects an overstep in a parse subroutine where a <= SER_LEN()
meant it would process the terminal as if it were a character and try
to run a case translation on it via LO_CASE().

(Note that in addition to being more rigorous and clear what's going
on to fix such cases, it also saves a small bit of processing!)