Restore DEBUG_UTF8_EVERYWHERE, add to Travis
While "UTF-8 Everywhere" was being implemented, a slower debugging
mode which utilized a separate set of C++ definitions was able to
catch programming errors related to the type difference between
*already validated* UTF-8 bytes and unvalidated ones.  It also helped
spot usage of simple "+ 1" style arithmetic to advance characters on
the validated UTF-8 bytes, enforcing the use of functions instead
and requiring explicit casts to do otherwise.

This mode had atrophied in the interim, so that type errors at
callsites in the C implementation were not getting checked.  This
commit brings the mode back up to function--extending it to support
new needs.

It also makes one of the Travis builds uses it, to keep violations of
the checks monitored.